### 2025-08-04 #8715 @editorialbot
[PRE REVIEW]: IdentityByDescentDispersal.jl: Inferring dispersal rates with identity-by-descent blocks
**Submitting author:** @currocam (<a href="http://orcid.org/0000-0001-8285-9318">Francisco Campuzano Jiménez</a>)
**Repository:** https://github.com/currocam/IdentityByDescentDispersal.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
