+++ .github/workflows/SnakemakeWorkflow.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout repository
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - name: Set up Julia
         uses: julia-actions/setup-julia@v2
