### 2025-09-01 238e932
chore: :robot: pre-commit update (#21)
Co-authored-by: currocam <69399781+currocam@users.noreply.github.com>
