+++ .pre-commit-config.yaml
@@ -8,7 +8,7 @@ repos:
         language: fail
         files: "\\.rej$"
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
+    rev: v6.0.0
     hooks:
       - id: check-json
       - id: check-toml
@@ -37,12 +37,12 @@ repos:
     hooks:
       - id: yamllint
   - repo: https://github.com/ericphanson/ExplicitImports.jl
-    rev: v1.13.1
+    rev: v1.13.2
     hooks:
       - id: explicit-imports
         name: ExplicitImports checks
         args: [--print, --checklist, exclude_all_qualified_accesses_are_public]
   - repo: https://github.com/domluna/JuliaFormatter.jl
-    rev: v2.1.5
+    rev: v2.1.6
     hooks:
       - id: julia-formatter
