+++ .github/workflows/Lint.yml
@@ -20,7 +20,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Clone
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
       - name: Setup Julia
         uses: julia-actions/setup-julia@v2
         with:
