+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"95-reference/#reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"95-reference/#Contents","page":"Reference","title":"Contents","text":"","category":"section"},{"location":"95-reference/","page":"Reference","title":"Reference","text":"Pages = [\"95-reference.md\"]","category":"page"},{"location":"95-reference/#Index","page":"Reference","title":"Index","text":"","category":"section"},{"location":"95-reference/","page":"Reference","title":"Reference","text":"Pages = [\"95-reference.md\"]","category":"page"},{"location":"95-reference/#IdentityByDescentDispersal.composite_loglikelihood_constant_density","page":"Reference","title":"IdentityByDescentDispersal.composite_loglikelihood_constant_density","text":"composite_loglikelihood_constant_density(D::Real, sigma::Real, df::DataFrame, contig_lengths::AbstractArray{<:Real}, chromosomal_edges::Bool=true, diploid::Bool=true) -> Real\n\nComputes the composite log-likelihood of the observed IBD blocks under a model with constant population density.\n\nD: Effective population density (diploid individuals per unit area).\nsigma: Root mean square dispersal distance per generation.\ndf: DataFrame containing the observed IBD blocks in the format returned by preprocess_dataset.\ncontig_lengths: Array of contig lengths in Morgans.\n\nOptionally:\n\nchromosomal_edges: Whether to account for chromosomal edge effects.\ndiploid: Whether to account for diploidy.\n\n\n\n\n\n","category":"function"},{"location":"95-reference/#IdentityByDescentDispersal.composite_loglikelihood_custom","page":"Reference","title":"IdentityByDescentDispersal.composite_loglikelihood_custom","text":"composite_loglikelihood_custom(De::Function, parameters::AbstractArray, sigma::Real, df::DataFrame, contig_lengths::AbstractArray{<:Real}, chromosomal_edges::Bool=true, diploid::Bool=true) -> Real\n\nComputes the composite log-likelihood of the observed IBD blocks under a model with constant population density.\n\nDe is  a user-defined function that takes time t and parameters and returns the effective population density at time t.\nparameters is a user-defined array of parameters that the function De depends on.\nsigma: Root mean square dispersal distance per generation.\ndf: DataFrame containing the observed IBD blocks in the format returned by preprocess_dataset.\ncontig_lengths: Array of contig lengths in Morgans.\n\nOptionally:\n\nchromosomal_edges: Whether to account for chromosomal edge effects.\ndiploid: Whether to account for diploidy.\n\n\n\n\n\n","category":"function"},{"location":"95-reference/#IdentityByDescentDispersal.composite_loglikelihood_power_density","page":"Reference","title":"IdentityByDescentDispersal.composite_loglikelihood_power_density","text":"composite_loglikelihood_power_density(D::Real, beta::Real, sigma::Real, df::DataFrame, contig_lengths::AbstractArray{<:Real}, chromosomal_edges::Bool=true, diploid::Bool=true) -> Real\n\nComputes the composite log-likelihood of the observed IBD blocks under a model with constant population density.\n\nD: Effective population density (diploid individuals per unit area).\nbeta is the power of the density function,\nsigma: Root mean square dispersal distance per generation.\ndf: DataFrame containing the observed IBD blocks in the format returned by preprocess_dataset.\ncontig_lengths: Array of contig lengths in Morgans.\n\nOptionally:\n\nchromosomal_edges: Whether to account for chromosomal edge effects.\ndiploid: Whether to account for diploidy.\n\n\n\n\n\n","category":"function"},{"
[…]
