### 2025-10-10 0b33a3c
Update docs & linter
