+++ docs/src/inference.jl
@@ -217,7 +217,7 @@ end
 # Such differences arise naturally when considering the effect of mating systems on dispersal patterns. For simplicity,
 # let's consider a 1-dimensional space where individuals can only move left or right as shown in the figure below.
 #
-# ![Illustration of dispersal in a 1-dimensional space with separate sexes](dispersal.svg){ width=100% }
+# ![Illustration of dispersal in a 1-dimensional space with separate sexes](dispersal.png)
 #
 # Let's assume the displacement between the mother and the offspring, $d_{\text{mother-child}}$, is distributed
 # according to a normal distribution with a mean of zero and a variance of $\sigma_D^2$ and
