+++ v0.1.3/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"95-reference/#reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"95-reference/#Contents","page":"Reference","title":"Contents","text":"","category":"section"},{"location":"95-reference/","page":"Reference","title":"Reference","text":"Pages = [\"95-reference.md\"]","category":"page"},{"location":"95-reference/#Index","page":"Reference","title":"Index","text":"","category":"section"},{"location":"95-reference/","page":"Reference","title":"Reference","text":"Pages = [\"95-reference.md\"]","category":"page"},{"location":"95-reference/#IdentityByDescentDispersal.age_density_ibd_blocks_custom","page":"Reference","title":"IdentityByDescentDispersal.age_density_ibd_blocks_custom","text":"age_density_ibd_blocks_custom(t::Real, r::Real, De::Function, parameters::AbstractArray, sigma::Real, L::Real, G::Real, chromosomal_edges::Bool = true, diploid::Bool = true)\n\nComputes the expected density of identity-by-descent (IBD) blocks of length L and age t for a model where the effective population density is given by a custom function De(t, parameters). This function returns the expected number of IBD blocks of age t per pair of individuals and per unit of block length.\n\nmathbbEN_L^t =  G  4t^2 exp(-2Lt) cdot Phi(t)\n\nwhere:\n\nt is the age of the IBD block,\nr is the geographic distance between samples,\nDe is a user-defined function that takes time t and a parameters and returns the effective population density at time t.\nparameters is a user-defined array of parameters that the function De depends on.\nsigma is the root mean square dispersal distance per generation,\nL is the length of the IBD block (in Morgans),\nG is the total map length of the genome (in Morgans),\n\nIf chromosomal_edges is true (the default), we account for chromosomal edge effects. If diploid is true, we multiply by a factor of 4 to account for the fact that each individual has two copies of each chromosome. For more details, see Appendix B.\n\nReference: Ringbauer, H., Coop, G., & Barton, N. H. (2017). Genetics, 205(3), 1335–1351.\n\n\n\n\n\n","category":"function"},{"location":"95-reference/#IdentityByDescentDispersal.composite_loglikelihood_constant_density","page":"Reference","title":"IdentityByDescentDispersal.composite_loglikelihood_constant_density","text":"composite_loglikelihood_constant_density(D::Real, sigma::Real, df::DataFrame, contig_lengths::AbstractArray{<:Real}, chromosomal_edges::Bool=true, diploid::Bool=true) -> Real\n\nComputes the composite log-likelihood of the observed IBD blocks under a model with constant population density.\n\nD: Effective population density (diploid individuals per unit area).\nsigma: Root mean square dispersal distance per generation.\ndf: DataFrame containing the observed IBD blocks in the format returned by preprocess_dataset.\ncontig_lengths: Array of contig lengths in Morgans.\n\nOptionally:\n\nchromosomal_edges: Whether to account for chromosomal edge effects.\ndiploid: Whether to account for diploidy.\n\n\n\n\n\n","category":"function"},{"location":"95-reference/#IdentityByDescentDispersal.composite_loglikelihood_custom","page":"Reference","title":"IdentityByDescentDispersal.composite_loglikelihood_custom","text":"composite_loglikelihood_custom(De::Function, parameters::AbstractArray, sigma::Real, df::DataFrame, contig_lengths::AbstractArray{<:Real}, chromosomal_edges::Bool=true, diploid::Bool=true) -> Real\n\nComputes the composite log-likelihood of the observed IBD blocks under a model with constant population density.\n\nDe is  a user-defined function that takes time t and parameters and returns the effective population density 
[…]
