+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
     hooks:
       - id: validate-cff
   - repo: https://github.com/google/yamlfmt
-    rev: v0.17.2
+    rev: v0.18.0
     hooks:
       - id: yamlfmt
   - repo: https://github.com/adrienverge/yamllint
