+++ .github/workflows/Draft.yml
@@ -15,7 +15,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: manuscript/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
