### 2025-11-11 9fef372
chore: :robot: pre-commit update (#26)
Co-authored-by: currocam <69399781+currocam@users.noreply.github.com>
