+++ .pre-commit-config.yaml
@@ -43,6 +43,6 @@ repos:
         name: ExplicitImports checks
         args: [--print, --checklist, exclude_all_qualified_accesses_are_public]
   - repo: https://github.com/domluna/JuliaFormatter.jl
-    rev: v2.1.6
+    rev: v2.2.0
     hooks:
       - id: julia-formatter
