### 2025-11-18 #8715 @currocam
Hi @espottesmith.

If I understood correctly, JOSS tries to prioritize authors of previous JOSS articles to review. In that case, and from the list of similar articles, I would first suggest tkchafin (`autoStreamTree: Genomic variant data fitted to geospatial networks`) as they have a relevant background in Julia development and spatial genetics.

From a more population genetics background, rivasiker (`PhaseTypeR: an R package for phase-type distributions in population genetics`) would also be a nice fit, perhaps also to consider the non-Julia user experience and review other aspects of the software such as the Snakemake pipeline or the SLiM simulation template. Reviewers of both articles would also be a good fit.
