### 2025-11-18 #8715 @tkchafin
I can review it Sent from my Galaxy
-------- Original message --------From: "Evan Walter Clark Spotte-Smith, PhD" ***@***.***> Date: 11/18/25  4:09 AM  (GMT-06:00) To: openjournals/joss-reviews ***@***.***> Cc: Tyler Chafin ***@***.***>, Mention ***@***.***> Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]:
  IdentityByDescentDispersal.jl: Inferring dispersal rates with
  identity-by-descent blocks (Issue #8715) espottesmith left a comment (openjournals/joss-reviews#8715)
@tkchafin @rivasiker @slwu89 @tpoisot @puruckertom
Would any of y'all be able to review this submission for JOSS?
