### 2025-11-22 c84ca93
chore: :robot: pre-commit update (#27)
Co-authored-by: currocam <69399781+currocam@users.noreply.github.com>
