+++ .pre-commit-config.yaml
@@ -37,7 +37,7 @@ repos:
     hooks:
       - id: yamllint
   - repo: https://github.com/ericphanson/ExplicitImports.jl
-    rev: v1.13.2
+    rev: v1.14.0
     hooks:
       - id: explicit-imports
         name: ExplicitImports checks
