### 2025-11-24 09ab036
Fix method overwriting during precompilation
