### 2025-11-24 2101b5c
Major simulations update 1. Before, I was using the correction factor from Chris C R Smith (2023). This, however, is not correct, as the dispersal we estimate is from a mixture of normals, rather than a normal itself. See documentation for the explanation. 2. I simplified the simulation to have reflecting boundaries and avoid dealing with the torus-surface logic. 3. Following the latest SLiM manual, I'm now using summarize to compute empirical densities. 4. One critical bug is that I was accidentally using steps from a truncated normal rather than a normal. Coincidentally, this error canceled 
[…]
