+++ simulations/simulate_constant_density_ibd_detection.jl
@@ -152,4 +152,4 @@ contig_lengths = [1.0]
 m = constant_density(df2, contig_lengths);
 mle_estimate = maximum_likelihood(m)
 coef_table = mle_estimate |> coeftable |> DataFrame
-pretty_table(coef_table, backend = Val(:markdown))
+pretty_table(coef_table)
