### 2025-11-24 691c12c
Rerun toy dataset & declare inline uv dependencies
