### 2025-11-24 60d2f4a
Use >1 threads in CI
