+++ docs/data/build.jl
@@ -1,0 +1,98 @@
+using PyCall, Random, DataFrames, StatsBase, PrettyTables
+using IdentityByDescentDispersal
+seed = 1000
+NE = 200 # Number of diploid individuals
+SD = 0.1 # Dispersal rate of the offspring
+SM = 0.01 # Mate choice kernel
+outpath = "s$(seed).trees"
+run(
+    `slim -p -s $seed -d NE=$NE -d SD=$SD -d SM=$SM -d OUTPATH="\"$outpath\"" ../../simulations/constant_density.slim`,
+)
+tskit = pyimport("tskit")
+ts = tskit.load(outpath);
+rng = MersenneTwister(seed)
+n_samples = 100
+sampled = randperm(rng, ts.num_individuals)[1:n_samples] .- 1
+nodes = reduce(vcat, [ts.individual(i).nodes for i in sampled])
+ts = ts.simplify(samples = nodes);
+function torus_distance(points::AbstractMatrix{<:Real})
+    coords = points[:, 1:2]
+    n = size(coords, 1)
+    dist_matrix = zeros(n, n)
+
+    for i = 1:n
+        for j = 1:n
+            dx = abs(coords[i, 1] - coords[j, 1])
+            dy = abs(coords[i, 2] - coords[j, 2])
+            dx = min(dx, 1 - dx)
+            dy = min(dy, 1 - dy)
+            dist_matrix[i, j] = sqrt(dx^2 + dy^2)
+        end
+    end
+    return dist_matrix
+end
+df_dist = let
+    points = ts.individual_locations
+    dist_matrix = torus_distance(points)
+    n = size(points, 1)
+    df = DataFrame(ID1 = Int[], ID2 = Int[], distance = Float64[])
+    for i = 1:n
+        for j = (i+1):n
+            push!(df, (i - 1, j - 1, dist_matrix[i, j]))  # 0-based ID
+        end
+    end
+    df
+end;
+df_ibds = let
+    function blocks(i, j)
+        ibds = ts.ibd_segments(
+            between = [ts.individual(i).nodes, ts.individual(j).nodes],
+            min_span = 1 / 100 * 4 * 1e8,
+            store_pairs = true,
+            store_segments = true,
+        )
+        spans = [
+            (block.right - block.left) / 1e8 for pair in ibds for block in ibds.get(pair)
+        ]
+        DataFrame(ID1 = i, ID2 = j, span = spans)
+    end
+    vcat([blocks(i, j) for i = 0:(n_samples-1) for j = (i+1):(n_samples-1)]...)
+end;
+
+# The package IdentityByDescentDispersal.jl has a function to preprocess both `DataFrames` into a single long DataFrame:
+# Here, we use the default bins used by Ringbauer et. al. in their paper.
+bins, min_threshold = default_ibd_bins()
+df_preprocessed = preprocess_dataset(df_ibds, df_dist, bins, min_threshold);
+
+# In order to speedup the inference we are going to also bin the distances:
+function cut(values::AbstractVector{<:Real}, edges::AbstractVector{<:Real})
+    n = length(values)
+    bins = Vector{Float64}(undef, n)
+    for i = 1:n
+        x = values[i]
+        bin_idx = findfirst(j -> edges[j] <= x < edges[j+1], 1:(length(edges)-1))
+        @assert !isnothing(bin_idx)
+        bins[i] = edges[bin_idx]  # label bin by left edge
+    end
+    return bins
+end
+
+df2 = let
+    edges_distance = collect(0:0.05:1.0)
[…]
