+++ docs/data/Project.toml
@@ -1,0 +1,7 @@
+[deps]
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+IdentityByDescentDispersal = "270ed24e-ac4d-5dce-b70a-bd124b3b694a"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
+PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
+PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
