### 2025-11-24 85aa3a6
Update manuscript to match improved simulated dataset
