+++ manuscript/paper.md
@@ -78,13 +78,13 @@ Both the SLiM simulation template and the Snakemake pipeline can be found in the
 
 # Example
 
-In this section, we demonstrate how `IdentityByDescentDispersal.jl` can be used together with the popular `Turing.jl` framework [@ge_turing_2018] using a dataset we simulate in the documentation. We analyze error-free IBD blocks shared by 100 diploid individuals from a constant-density population with parameters $D_{\text{true}}\approx 250$ diploids/km² and $\sigma_{\text{true}}\approx 0.071$ km/generation.
+In this section, we demonstrate how `IdentityByDescentDispersal.jl` can be used together with the popular `Turing.jl` framework [@ge_turing_2018] using a dataset we simulate in the documentation. We analyze error-free IBD blocks shared by 100 diploid individuals from a constant-density population with parameters $D_{\text{true}}\approx 200$ diploids/km² and $\sigma_{\text{true}}\approx 0.100$ km/generation.
 
 `IdentityByDescentDispersal.jl` has extensive documentation that covers the underlying theory behind the method, how to effectively simulate synthetic datasets, various demographic models, and inference algorithms. We refer the reader to the documentation for more details, which can be found at [https://currocam.github.io/IdentityByDescentDispersal.jl/](https://currocam.github.io/IdentityByDescentDispersal.jl/dev/).
 
 Thanks to `Turing.jl`, we can perform Bayesian inference with a wide range of popular Monte Carlo algorithms. \autoref{fig:example} shows the estimated pseudo-posterior obtained through doing inference with the composite likelihood.
 
-![Estimated pseudo-posterior obtained by doing inference with the composite likelihood. Although the pseudo-posterior is not well calibrated, it concentrates near the true values ($\mathbb E[D | \text{data}]\approx 281$ and $\mathbb E[\sigma | \text{data}]\approx 0.068$, respectively).\label{fig:example}](figures/nuts_constant_density.svg){ width=100% }
+![Estimated pseudo-posterior obtained by doing inference with the composite likelihood. The pseudo-posterior concentrates around the true values ($\mathbb E[D | \text{data}]\approx 200$ and $\mathbb E[\sigma | \text{data}]\approx 0.102$, respectively). However, we don’t generally expect to be well-calibrated, and we suggest performing simulation-based calibration checking.\label{fig:example}](figures/nuts_constant_density.svg){ width=100% }
 
 \autoref{fig:example} was generated by the following snippet of Julia code, which reads the processed data CSV from the provided Snakemake pipeline.
 
@@ -113,7 +113,7 @@ mle_estimate = maximum_likelihood(
 coeftable(mle_estimate)
 ```
 
-which estimates $D_{\text{MLE}}\approx 282$ diploids/km² (95% CI: 260–303) and $\sigma_{\text{MLE}}\approx 0.068$ km/generation (95% CI: 0.065–0.071). The 95% confidence interval is computed from the Fisher information matrix.
+which estimates $D_{\text{MLE}}\approx 199$ diploids/km² (95% CI: 171–226) and $\sigma_{\text{MLE}}\approx 0.102$ km/generation (95% CI: 0.094–0.110). The 95% confidence interval is computed from the Fisher information matrix.
 
 # Availability
 
