### 2025-11-25 #8715 @currocam
Hi, I'm going to regenerate the PDF to match the latest changes.

(1) A couple of months ago, I was contacted by some researchers asking if a figure such as [this posterior predictive plot](https://currocam.github.io/IdentityByDescentDispersal.jl/dev/inference/f9f41171.svg) was possible to make with the package. It was not, so I added it and [documented it](https://currocam.github.io/IdentityByDescentDispersal.jl/dev/inference/f9f41171.svg). I accidentally meshed up with the package precompilation, but it should be solved now.

(2) Revising the material, I found two bugs in the SLiM code I was using to simulate the synthetic dataset I used in the documentation and example figure. There was a factor of 2 in an equation that somehow canceled out calling a function with the wrong argument order.
It fixed it in [here](https://github.com/currocam/IdentityByDescentDispersal.jl/commit/2101b5cb9116f3d39a8c54eeb2f23bae94235443). I can elaborate if needed, but I'm confident the current state is ready to review :)
