### 2025-11-25 #8715 @espottesmith
@Treys925 @chenyangkang @g-rppl @gchure @GJHSimmons

can any of y'all contribute your expertise to review this manuscript/codebase?
