### 2025-11-25 #8715 @chenyangkang
>
> can any of y'all contribute your expertise to review this manuscript/codebase?

I'm not familiar with Julia, unfortunately.
