+++ versions.js
@@ -3,5 +3,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.1.3";
+var DOCUMENTER_NEWEST = "0.1.4";
 var DOCUMENTER_STABLE = "stable";
