### 2025-12-03 496386f
Fallback to quadrature rule for large bins
