### 2026-01-16 fe00231
Improved nested integration
