+++ Project.toml
@@ -1,21 +1,23 @@
 name = "IdentityByDescentDispersal"
 uuid = "270ed24e-ac4d-5dce-b70a-bd124b3b694a"
-authors = ["Curro Campuzano Jiménez campuzanocurro@gmail.com"]
 version = "0.1.5"
+authors = ["Curro Campuzano Jiménez campuzanocurro@gmail.com"]
 
 [deps]
 BesselK = "432ab697-7a72-484f-bc4a-bc531f5c819b"
 DataAPI = "9a962f9c-6df0-11e9-0e5d-c546b8b5ee8a"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+HCubature = "19dc6840-f33b-545b-b366-655c7e3ffd49"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
-QuadGK = "2"
-Distributions = "0.25"
-Tables = "1"
-DataFrames = "1"
-DataAPI = "1"
 BesselK = "0.5"
+DataAPI = "1"
+DataFrames = "1"
+Distributions = "0.25"
+HCubature = "1.8.0"
+QuadGK = "2"
+Tables = "1"
 julia = "1.10"
