### 2026-01-21 6568ff7
Fix distance function error and improve simulation study cases
