+++ simulations/simulate_habitat_suitability_ground_truth.jl
@@ -1,0 +1,229 @@
+# # Habitat suitability simulation
+using PyCall, Random, DataFrames, StatsBase, PrettyTables
+using IdentityByDescentDispersal
+using Turing
+# ## Forward-in-time simulation
+
+run(`slim --version`)
+
+function euclidean_distance(points)
+    coords = points[:, 1:2]
+    n = size(coords, 1)
+    dist_matrix = zeros(n, n)
+    for i = 1:n
+        for j = 1:n
+            dx = coords[i, 1] - coords[j, 1]
+            dy = coords[i, 2] - coords[j, 2]
+            dist_matrix[i, j] = sqrt(dx^2 + dy^2)
+        end
+    end
+    return dist_matrix
+end
+# In order to speedup the inference we are going to also bin the distances:
+function cut(values::AbstractVector{<:Real}, edges::AbstractVector{<:Real})
+    n = length(values)
+    bins = Vector{Float64}(undef, n)
+    for i = 1:n
+        x = values[i]
+        bin_idx = findfirst(j -> edges[j] <= x < edges[j+1], 1:(length(edges)-1))
+        @assert !isnothing(bin_idx)
+        bins[i] = edges[bin_idx]  # label bin by left edge
+    end
+    return bins
+end
+
+# Define the simulation loop
+function simulation(NE, SD, SM, seed)
+    outpath = "habitat_suitability.trees"
+    run(
+        `slim -s $seed -d NE=$NE -d SD=$SD -d SM=$SM -d OUTPATH="\"$outpath\"" habitat_suitability.slim`,
+    )
+    # ## Data preprocessing
+    # We analyze the simulated tree-sequence using the python library `tskit`. Because IBD blocks decay very quickly with time and
+    # we have run the simulation for many generations (5000), we don't need to perform `recapitation`.
+    tskit = pyimport("tskit")
+    ts = tskit.load(outpath);
+
+    # We take a sample of 200 diploid individuals.
+    rng = MersenneTwister(seed)
+    n_samples = 200
+    sampled = randperm(rng, ts.num_individuals)[1:n_samples] .- 1
+    nodes = reduce(vcat, [ts.individual(i).nodes for i in sampled])
+    ts = ts.simplify(samples = nodes);
+
+    df_dist = let
+        points = reduce(hcat, [collect(row) for row in ts.individual_locations])'
+        dist_matrix = euclidean_distance(points)
+        n = size(points, 1)
+        df = DataFrame(ID1 = Int[], ID2 = Int[], distance = Float64[])
+        for i = 1:n
+            for j = (i+1):n
+                push!(df, (i - 1, j - 1, dist_matrix[i, j]))  # 0-based ID
+            end
+        end
+        df
+    end;
+
+    # And the IBD blocks:
+    df_ibds = let
+        function blocks(i, j)
+            ibds = ts.ibd_segments(
+                between = [ts.individual(i).nodes, ts.individual(j).nodes],
+                min_span = 1 / 100 * 4 * 1e8,
+                store_pairs = true,
+                store_segments = true,
+            )
+            spans = [
+                (block.right - block.left) / 1e8 for pair in ibds for
+                block in ibds.get(pair)
+            ]
+            DataFrame(ID1 = i, ID2 = j, span = spans)
+        end
+        vcat([blocks(i, j) for i = 0:(n_samples-1) for j = (i+1):(n_samples-1)]...)
+    end;
+
+    # The package IdentityByDescentDispersal.jl has a function to preprocess both `DataFrames` into a single long DataFrame:
[…]
