### 2026-01-21 97432fd
Improve documentation and fix building CI
