### 2026-01-22 657ebc0
Dev (#34)
* Improved nested integration

* Fix distance function error and improve simulation study cases

* Improve documentation and fix building CI
