+++ simulations/simulate_constant_density_ground_truth.jl
@@ -2,7 +2,6 @@
 using PyCall, Random, DataFrames, StatsBase, PrettyTables
 using IdentityByDescentDispersal
 # ## Forward-in-time simulation
-# We will use SLiM to simulate a constant density population living in a 2D torus and `tskit` and tree-sequence recording to analyze the ground truth of IBD blocks.
 
 run(`slim --version`)
 
@@ -14,6 +13,19 @@ seed = 1000
 NE = 200 # Number of diploid individuals
 SD = 0.1 # Dispersal rate of the offspring
 SM = 0.01 # Mate choice kernel
+function euclidean_distance(points)
+    coords = points[:, 1:2]
+    n = size(coords, 1)
+    dist_matrix = zeros(n, n)
+    for i = 1:n
+        for j = 1:n
+            dx = coords[i, 1] - coords[j, 1]
+            dy = coords[i, 2] - coords[j, 2]
+            dist_matrix[i, j] = sqrt(dx^2 + dy^2)
+        end
+    end
+    return dist_matrix
+end
 outpath = "s$(seed).trees"
 run(
     `slim -p -s $seed -d NE=$NE -d SD=$SD -d SM=$SM -d OUTPATH="\"$outpath\"" constant_density.slim`,
@@ -32,26 +44,9 @@ sampled = randperm(rng, ts.num_individuals)[1:n_samples] .- 1
 nodes = reduce(vcat, [ts.individual(i).nodes for i in sampled])
 ts = ts.simplify(samples = nodes);
 
-# Then, we compute pairwise distances across individuals (in the torus)
-function torus_distance(points::AbstractMatrix{<:Real})
-    coords = points[:, 1:2]
-    n = size(coords, 1)
-    dist_matrix = zeros(n, n)
-
-    for i = 1:n
-        for j = 1:n
-            dx = abs(coords[i, 1] - coords[j, 1])
-            dy = abs(coords[i, 2] - coords[j, 2])
-            dx = min(dx, 1 - dx)
-            dy = min(dy, 1 - dy)
-            dist_matrix[i, j] = sqrt(dx^2 + dy^2)
-        end
-    end
-    return dist_matrix
-end
 df_dist = let
-    points = ts.individual_locations
-    dist_matrix = torus_distance(points)
+    points = reduce(hcat, [collect(row) for row in ts.individual_locations])'
+    dist_matrix = euclidean_distance(points)
     n = size(points, 1)
     df = DataFrame(ID1 = Int[], ID2 = Int[], distance = Float64[])
     for i = 1:n
@@ -120,7 +115,17 @@ using Turing
 @model function constant_density(df, contig_lengths)
     D ~ Uniform(0, 4000)
     σ ~ Uniform(0, 1)
-    Turing.@addlogprob! composite_loglikelihood_constant_density(D, σ, df, contig_lengths)
+    try
+        Turing.@addlogprob! composite_loglikelihood_constant_density(
+            D,
+            σ,
+            df,
+            contig_lengths,
+        )
+    catch e
+        @warn "Error in constant_density model: $e"
+        Turing.@addlogprob! -Inf
+    end
 end
 contig_lengths = [1.0]
 m = constant_density(df2, contig_lengths);
