+++ simulations/simulate_constant_density_ibd_detection.md
@@ -10,7 +10,6 @@ using IdentityByDescentDispersal
 ````
 
 ## Forward-in-time simulation
-We will use SLiM to simulate a constant density population living in a 2D torus and `tskit` and tree-sequence recording to analyze the ground truth of IBD blocks.
 
 ````julia
 run(`slim --version`)
@@ -30,6 +29,19 @@ seed = 1000
 NE = 200 # Number of individuals
 SD = 0.1 # Dispersal rate of the offspring
 SM = 0.01 # Mate choice kernel
+function euclidean_distance(points)
+    coords = points[:, 1:2]
+    n = size(coords, 1)
+    dist_matrix = zeros(n, n)
+    for i = 1:n
+        for j = 1:n
+            dx = coords[i, 1] - coords[j, 1]
+            dy = coords[i, 2] - coords[j, 2]
+            dist_matrix[i, j] = sqrt(dx^2 + dy^2)
+        end
+    end
+    return dist_matrix
+end
 outpath = "s$(seed).trees"
 run(
     `slim -s $seed -d NE=$NE -d SD=$SD -d SM=$SM -d OUTPATH="\"$outpath\"" constant_density.slim`,
@@ -53,6 +65,12 @@ We perform recapitation to ensure samples are fully coalesced
 ````julia
 pyslim = pyimport("pyslim")
 rts = pyslim.recapitate(ts, ancestral_Ne = NE, recombination_rate = 1e-8);
+````
+
+````
+/Users/curro/.julia/conda/3/aarch64/lib/python3.12/site-packages/msprime/ancestry.py:1290: TimeUnitsMismatchWarning: The initial_state has time_units=ticks but time is measured in generations in msprime. This may lead to significant discrepancies between the timescales. If you wish to suppress this warning, you can use, e.g., warnings.simplefilter('ignore', msprime.TimeUnitsMismatchWarning)
+  sim = _parse_sim_ancestry(
+
 ````
 
 We take a sample of 100 diploid individuals.
@@ -96,7 +114,7 @@ run(`curl -o hap-ibd.jar https://faculty.washington.edu/browning/hap-ibd.jar`)
 Process(`curl -o hap-ibd.jar https://faculty.washington.edu/browning/hap-ibd.jar`, ProcessExited(0))
 ````
 
-Create a dummy genetic map in Plink format
+Create a dummy genetic map in Plink format for recombination rate 1e-8
 
 ````julia
 mapfile = "s$(seed).plink.map"
@@ -108,11 +126,11 @@ end
 Execute HapIBD software
 
 ````julia
-run(`java -jar hap-ibd.jar gt=$outvcf map=$mapfile out=s$(seed)`)
+run(`java -Duser.language=en -Duser.country=US -jar hap-ibd.jar gt=$outvcf map=$mapfile out=s$(seed)`)
 ````
 
 ````
-Process(`java -jar hap-ibd.jar gt=s1000.vcf map=s1000.plink.map out=s1000`, ProcessExited(0))
+Process(`java -Duser.language=en -Duser.country=US -jar hap-ibd.jar gt=s1000.vcf map=s1000.plink.map out=s1000`, ProcessExited(0))
 ````
 
 Preprocess detected IBD blocks
@@ -123,44 +141,29 @@ run(
 )
 postprocessed_file = "s$(seed).postprocessed.ibd"
 run(
-    `bash -c "gunzip -c s$(seed).ibd.gz | java -jar merge-ibd-segments.jar $outvcf $mapfile 0.6 1 > $postprocessed_file"`,
[…]

[…]
