+++ dev/tutorial.jl
@@ -14,6 +14,7 @@ using IdentityByDescentDispersal
 # Next, we show how to compute the expected density of identity-by-descent blocks under a constant population density using the `expected_ibd_blocks_constant_density` function.
 
 using Plots
+using QuadGK
 
 L = 0.01         # Block length threshold (in Morgans)
 G = 1.0          # Genome length (in Morgans)
@@ -56,6 +57,10 @@ plot!(
 # individuals that are 1 kM apart in a population with constant effective population density of `10/kM²`and dispersal rate of `0.5/kM` would be
 
 expected_ibd_blocks_constant_density(1.0, 10.0, 0.5, L, G) * 0.001
+
+# However, we can also compute the expected number of blocks for any bin size using numerical integration (and we do this when computing log-likelihoods internally when necessary):
+
+quadgk(x -> expected_ibd_blocks_constant_density(x, 10.0, 0.5, L, G) * 0.001, 1.0, 1.1)[1]
 
 # Ringbauer et al. (2017) derived analytical solutions for the expected density of identity-by-descent blocks for
 # the family of functions of effective population density of the form $D_e(t) = D_1 t^{-\beta}$. A constant effective population density is a special
