### 2026-02-08 #9517 @slwu89
@espottesmith I've made 2 issues in the repository for my review comments.

Regarding "Contribution and authorship: Has the submitting author (@currocam) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?" the first 2 authors on the paper have clearly contributed to the software. However the remaining authors have not made any commits to the repository. I suggest either to add a section to the paper describing why they should be included despite lack of software contributions, or the author list should be revised. I would like to know your opinion here as editor, I am not sure if JOSS includes author contributions list as a standard section.
