### 2026-02-09 #9517 @espottesmith
JOSS does not require an author contribution statement, though it's also not specifically disallowed/discouraged.

From our submission guidelines:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

So, if the non-code-contributing individuals intellectually contributed to the software design or to the direction of software development, they can certainly be listed as authors. Their contributions don't have to be explained in the manuscript, but it's legitimate to ask for @currocam to explain everyone's contributions.
