### 2026-02-09 #9517 @currocam
Thanks for the feedback! I'll work on it this week.

@slwu89  @espottesmith About the authorship : I am the main developer of the package (which I started porting from an unpublished working prototype I made). Arthur Zwaenepoel (@arzwa) took part in the code revision of the final package and in writing large parts of the technical documentation. Els De Keyzer gave feedback on the documentation and provided guidance in the development of the bioinformatics pipeline (i.e. the Snakemake workflow), which we also share together with the project for the pre-processing steps to make the necessary dataset. Hannes Svardal (@feilchenfeldt) made conceptual contributions and provided guidance in the development of the initial prototype and designing the simulation setup, which is a secondary part of the project.

My understanding from JOSS guidelines is that both Hannes Svardal and Els De Keyzer’s participation falls into the category “active project direction and other forms of non-code contributions”. But we are happy to revise the authorship if you think otherwise.
