### 2026-02-09 #9517 @espottesmith
From what you've described, your author list as it stands seems legitimate and valid.
