+++ manuscript/paper.bib
@@ -264,3 +264,22 @@
   year = {2013},
   howpublished = {\url{https://github.com/JuliaMath/QuadGK.jl}}
 }
+@article {osmond,
+article_type = {journal},
+title = {Estimating dispersal rates and locating genetic ancestors with genome-wide genealogies},
+author = {Osmond, Matthew and Coop, Graham},
+editor = {Nordborg, Magnus and Weigel, Detlef and Nordborg, Magnus and Speidel, Leo and Kelleher, Jerome},
+volume = 13,
+year = 2024,
+month = {nov},
+pub_date = {2024-11-26},
+pages = {e72177},
+citation = {eLife 2024;13:e72177},
+doi = {10.7554/eLife.72177},
+url = {https://doi.org/10.7554/eLife.72177},
+abstract = {Spatial patterns in genetic diversity are shaped by individuals dispersing from their parents and larger-scale population movements. It has long been appreciated that these patterns of movement shape the underlying genealogies along the genome leading to geographic patterns of isolation-by-distance in contemporary population genetic data. However, extracting the enormous amount of information contained in genealogies along recombining sequences has, until recently, not been computationally feasible. Here, we capitalize on important recent advances in genome-wide gene-genealogy reconstruction and develop methods to use thousands of trees to estimate per-generation dispersal rates and to locate the genetic ancestors of a sample back through time. We take a likelihood approach in continuous space using a simple approximate model (branching Brownian motion) as our prior distribution of spatial genealogies. After testing our method with simulations we apply it to \textit{Arabidopsis thaliana}. We estimate a dispersal rate of roughly 60 km\textsuperscript{2}/generation, slightly higher across latitude than across longitude, potentially reflecting a northward post-glacial expansion. Locating ancestors allows us to visualize major geographic movements, alternative geographic histories, and admixture. Our method highlights the huge amount of information about past dispersal events and population movements contained in genome-wide genealogies.},
+keywords = {ancestral recombination graph, tree sequence, spatial inference, branching Brownian motion, population genetics, phylogeography, importance sampling},
+journal = {eLife},
+issn = {2050-084X},
+publisher = {eLife Sciences Publications, Ltd},
+}
