### 2026-02-09 0dd4427
Add references and improve text
