### 2026-02-09 e316991
Safe likelihood calculation
