+++ src/IdentityByDescentDispersal.jl
@@ -412,6 +412,7 @@ Computes the composite log-likelihood of the observed IBD blocks under a model w
 Optionally:
 - `chromosomal_edges`: Whether to account for chromosomal edge effects.
 - `diploid`: Whether to account for diploidy.
+- `verbose`: Whether to print warnings when the computation fails.
 
 """
 function composite_loglikelihood_constant_density(
@@ -421,62 +422,68 @@ function composite_loglikelihood_constant_density(
     contig_lengths::AbstractArray{<:Real};
     chromosomal_edges::Bool = true,
     diploid::Bool = true,
+    verbose::Bool = true
 )
     # Input validation
     if D ≤ 0 || sigma ≤ 0 || isempty(contig_lengths)
         return -Inf
     end
+    try
+        log_likelihood = 0.0
+        for row in eachrow(df)
+            r = row.DISTANCE
+            L_left = row.IBD_LEFT
+            L_right = row.IBD_RIGHT
+            nr_pairs = row.NR_PAIRS
+            count = row.COUNT
+            ΔL = L_right - L_left
 
-    log_likelihood = 0.0
-
-    for row in eachrow(df)
-        r = row.DISTANCE
-        L_left = row.IBD_LEFT
-        L_right = row.IBD_RIGHT
-        nr_pairs = row.NR_PAIRS
-        count = row.COUNT
-        ΔL = L_right - L_left
-
-        # Compute expected IBD blocks
-        if ΔL < BIN_THRESHOLD
-            λ =
-                expected_ibd_blocks_constant_density(
-                    r,
-                    D,
-                    sigma,
+            # Compute expected IBD blocks
+            if ΔL < BIN_THRESHOLD
+                λ =
+                    expected_ibd_blocks_constant_density(
+                        r,
+                        D,
+                        sigma,
+                        L_left,
+                        contig_lengths,
+                        chromosomal_edges,
+                        diploid,
+                    ) *
+                    ΔL *
+                    nr_pairs
+            else
+                λ, _ = quadgk(
+                    x -> expected_ibd_blocks_constant_density(
+                        r,
+                        D,
+                        sigma,
+                        x,
+                        contig_lengths,
+                        chromosomal_edges,
+                        diploid,
+                    ),
                     L_left,
-                    contig_lengths,
-                    chromosomal_edges,
-                    diploid,
-                ) *
-                ΔL *
-                nr_pairs
-        else
-            λ, _ = quadgk(
-                x -> expected_ibd_blocks_constant_density(
-                    r,
-                    D,
-                    sigma,
-                    x,
-                    contig_lengths,
-                    chromosomal_edges,
-                    diploid,
-                ),
-                L_left,
-                L_right,
-            )
-            λ *= nr_pairs
-        end
+                    L_right,
+                )
+                λ *= nr_pairs
+            end
 
-        # Check for invalid λ
-        if λ < 0 || isnan(λ)
-            return -Inf
-        end
+            # Check for invalid λ
+            if λ < 0 || isnan(λ)
+                return -Inf
+            end
 
-        log_likelihood += logpdf(Poisson(λ), count)
+            log_likelihood += logpdf(Poisson(λ), count)
+        end
+        return log_likelihood
+    catch e
+        if verbose
+            @warn("(De=$(D), σ=$(sigma)) failed with error: $(e)")
+        end
+        return -Inf
     end
 
-    return log_likelihood
 end
 
 """
@@ -492,6 +499,7 @@ Computes the composite log-likelihood of the observed IBD blocks under a model w
 Optionally:
 - `chromosomal_edges`: Whether to account for chromosomal edge effects.
 - `diploid`: Whether to account for diploidy.
+- `verbose`: Whether to print warnings when the computation fails.
 
 """
[…]
