+++ manuscript/paper.md
@@ -48,7 +48,7 @@ The method of @ringbauer_inferring_2017 was limited to a family of functions for
 
 Lastly, our package comes with a template to simulate synthetic datasets and a pipeline for end-to-end analysis from VCF files to final estimates. We believe it will encourage a broader audience to adopt the inference scheme proposed by @ringbauer_inferring_2017, motivate further developments and expand its applications.
 
-Other related software includes `spacetrees` [@osmond] which estimates dispersal rates from inferred genome-wide genealogies and `disperseNN2` [@smith_dispersenn2_2023], a machine learning framework for predicting the displacement distance from genetic variation data. `IdentityByDescentDispersal.jl` differs in the nature of the data it uses for inference, but also in that it allows for flexible model-based inference of the effective density. Moreover, we expect `IdentityByDescentDispersal.jl` to be several orders of magnitude faster than `disperseNN2` for large datasets. For example, finding the MLE of the example dataset (n=100 diploids) takes less than a minute in a single CPU with `IdentityByDescentDispersal.jl`, whereas @smith_dispersenn2_2023 reported an approximate runtime of a week with a GPU training the disperseNN2 model for n=100 diploids.
+Other related software includes `spacetrees` [@osmond], which estimates dispersal rates from inferred genome-wide genealogies, and `disperseNN2` [@smith_dispersenn2_2023], a machine learning framework for predicting the expected per-generation displacement distance. `IdentityByDescentDispersal.jl` differs in the nature of the data it uses for inference, but also in that it allows for flexible model-based inference of the effective density. Moreover, we expect `IdentityByDescentDispersal.jl` to be several orders of magnitude faster than `disperseNN2` for large datasets. For example, @smith_dispersenn2_2023 reported an approximate runtime of a week to train the disperseNN2 model with a GPU for n=100 diploids, whereas using `IdentityByDescentDispersal.jl`, finding the MLE of the example dataset with n=100 diploids takes less than a minute in a single CPU.
 
 # Overview
 
