### 2026-02-09 98e08d4
Review feedback (#37)
* Add related software with runtime mention

* Solves: `Community guidelines (e.g. how to get support, etc) are not clearly mentioned in the README.md.

* Add references and improve text

* Safe likelihood calculation

* Fix linter complains

* Fix author format

* Improve documentation

* Better?
