+++ src/IdentityByDescentDispersal.jl
@@ -1,11 +1,9 @@
 module IdentityByDescentDispersal
 using BesselK: adbesselk
 using QuadGK: quadgk
-using DataFrames: DataFrame, transform
+using DataFrames: DataFrame, transform, ByRow, nrow
 using Distributions: Poisson, logpdf
 using HCubature: hcubature
-import Tables
-import DataAPI
 export safe_adbesselk,
     expected_ibd_blocks_constant_density,
     expected_ibd_blocks_power_density,
@@ -71,25 +69,22 @@ end
 Computes the expected density of identity-by-descent (IBD) blocks of length `L` for a model with constant population density.
 This function returns the expected number of IBD blocks per pair of individuals and per unit of block length.
 
+- `r` is the geographic distance between samples,
+- `D` is the effective population density (diploid individuals per unit area),
+- `sigma` is the root mean square dispersal distance per generation,
+- `L` is the length of the IBD block (in Morgans),
+- `G` is the total map length of the genome (in Morgans)
+
+If `chromosomal_edges` is `true` (the default), we account for chromosomal edge effects. If `diploid` is `true`, we multiply by a factor of 4 to account for the fact that each individual has two copies of each chromosome. There is a function overload that accepts a vector of `G` values and returns the aggregated expected number of IBD blocks.
+
+We calculate it as
 ```math
 \\mathbb{E}[N_L] = \\int_0^\\infty \\mathbb{E}[N_L^t] \\,dt =
 \\frac{G}{8\\pi D \\sigma^2}
 \\left(\\frac{r}{\\sqrt{L} \\sigma}\\right)^2
 K_2\\left(\\frac{\\sqrt{2L} \\, r}{\\sigma}\\right)
 ```
-
-where:
-- `r` is the geographic distance between samples,
-- `D` is the effective population density (diploid individuals per unit area),
-- `sigma` is the root mean square dispersal distance per generation,
-- `L` is the length of the IBD block (in Morgans),
-- `G` is the total map length of the genome (in Morgans),
-- `K₂` is the modified Bessel function of the second kind of order 2.
-
-If `chromosomal_edges` is `true` (the default), we account for chromosomal edge effects. If `diploid` is `true`, we multiply by a factor of 4 to account for the fact that each individual has two copies of each chromosome. For more details, see Appendix B.
-
-There is a function overload that accepts a vector of `G` values and returns the aggregated expected number of IBD blocks.
-
+where `K₂` is the modified Bessel function of the second kind of order 2.
 Reference:
 Ringbauer, H., Coop, G., & Barton, N. H. (2017). Genetics, 205(3), 1335–1351.
 """
@@ -145,6 +140,17 @@ end
 
 Computes the expected density of identity-by-descent (IBD) blocks of length `L` for a model with a power population density in the form of ``D(t) = D_0t^{-\beta}``
 This function returns the expected number of IBD blocks per pair of individuals and per unit of block length.
+
+- `r` is the geographic distance between samples,
+- `D` is the effective population density (diploid individuals per unit area),
+- `beta` is the power of the density function,
+- `sigma` is the root mean square dispersal distance per generation,
+- `L` is the length of the IBD block (in Morgans),
+- `G` is the total map length of the genome (in Morgans),
+
+If `chromosomal_edges` is `true` (the default), we account for chromosomal edge effects. If `diploid` is `true`, we multiply by a factor of 4 to account for the fact that each individual has two copies of each chromosome. There is a function overload that accepts a vector of `G` values and returns the aggregated expected number of IBD blocks.
+
+We calculate it as
 ```math
 \\mathbb{E}[N_L] = \\int_0^\\infty \\mathbb{E}[N_L^t] \\,dt =
 2^{\\frac{-3\\beta}{2}-3}
[…]
