### 2026-02-11 #9517 @g-rppl
@espottesmith, this seems to be a very useful package, well done @currocam et al.!
I am not a genomics person and can not comment much on the internal methods.
However, I really like how the provided functions integrate with the Julia ecosystem -- especially Turing.jl.
Most issues I found were also raised by @slwu89 and have already been addressed by the authors.

- [x] Although the paper does a good job in explaining the package's scope, workflow, and how it integrates with the Julia ecosystem, I struggled with these points looking at the documentation. Maybe adding an brief 'Overview' or 'About' section to the main documentation page can help users to get started.
- [x] Maybe also the 'Overview' section in the README can be streamlines a bit. The current version highlights the Snakemake pipeline, but its not clear to me what the package it self is capable of.
