### 2026-02-11 04ee866
Improve docs & simplify tutorial so CI runs fast enough for GitHub
