+++ simulations/README.md
@@ -1,12 +1,24 @@
-In addition to tests that assert the correctness of the implementation, we also include the necessary code to perform continuous-in-space simulations using SLiM and tree-sequence recording. Their goal is to simulate realistic datasets to test the method.
+# Simulations
 
-To this date, it is not clear in which regimes the approximation proposed by Ringbauer breaks, nor the expected amount of error introduced when detecting IBD blocks from phased data. Researchers should simulate data as part of their research workflow that matches their system to adapt their expectations.
+To this date, it is not clear in which regimes the approximation proposed by Ringbauer breaks, nor the expected amount of error introduced when detecting IBD blocks from phased data. We advise researchers to simulate data as part of their research workflow.
+
+To facilitate this, we provide a set of simulations of population genetics in a continuous space that use forward-in-time simulations (with SLiM), tree-sequence recording, and coalescent simulation (Haller et al., 2019; Baumdicker et al., 2022; Haller & Messer, 2023).
+
+## Simulation-based calibration
+
+Simulation-based calibration (using power-composite likelihoods) is an unexplored and interesting approach. Feel free to contact me if you are interested in this topic.
+
+## Basic recipes
 
 We include two basic scripts that can be adapted by others to match their case.
 
 - [Compute MLE estimate of a constant population with the ground truth IBD blocks](simulate_constant_density_ground_truth.md)
 - [Compute MLE estimate of a constant population with IBD blocks detected with HapIBD and post-processed with merge-ibd.jar](simulate_constant_density_ibd_detection.md)
 
-## Habitat suitability
+## Advanced recipes (WIP)
+
+Additionally, we provide an (incomplete) set of more advanced recipes.
+
+### Habitat suitability
 
 In practice, the assumption that individuals disperse according to a Brownian walk is not a modelling choice, but rather a statistical requirement for tractability. It is of great interest to the practitioner to address how such deviations from the model assumptions affect the inference. The [simulate_habitat_suitability_ground_truth.jl](simulate_habitat_suitability_ground_truth.jl) file can be used or adapted to study the effect of habitat patches on the accuracy of the inference.
