+++ docs/src/95-reference.md
@@ -1,4 +1,4 @@
-# [Reference](@id reference)
+# API reference
 
 ## Contents
 
