### 2026-02-11 a0b971d
Update README.md (#38)
* Update README.md

* Manually merge

---------

Co-authored-by: Curro Campuzano <69399781+currocam@users.noreply.github.com>
