### 2026-02-12 #9517 @currocam
Hello!
Thanks both of you for the feedback.

- We are tracking issues related to code in [JOSS Review Comments: Package only](https://github.com/currocam/IdentityByDescentDispersal.jl/issues/35) and they seem to be solved.
- @slwu89 pointed out in [JOSS Review Comments: Paper](https://github.com/currocam/IdentityByDescentDispersal.jl/issues/36) that the manuscript didn’t have enough detail on similar packages. I’ve added a new paragraph with two relevant examples and how they compared to this software.
- I agree with @g-rppl that the README was not very clear. I think the current [README](https://github.com/currocam/IdentityByDescentDispersal.jl) is now more concise and gives a better overview. More extensive documentation is in the documentation website.
- As suggested by @g-rppl, the [homepage](https://currocam.github.io/IdentityByDescentDispersal.jl/dev/) now has a high-level description of the package and what to expect in each section.
