### 2026-02-13 64a88bf
Add CITATION.cff
