+++ CITATION.cff
@@ -1,0 +1,27 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+type: software
+authors:
+  - given-names: Francisco
+    family-names: Campuzano Jiménez
+    affiliation: 'University of Antwerp, Belgium'
+    orcid: 'https://orcid.org/0000-0001-8285-9318'
+    email: curro.campuzanojimenez@uantwerpen.be
+  - given-names: Arthur
+    family-names: Zwaenepoel
+    affiliation: 'University of Antwerp, Belgium'
+    orcid: 'https://orcid.org/0000-0003-1085-2912'
+  - given-names: 'Els Lea R '
+    family-names: De Keyzer
+    orcid: 'https://orcid.org/0000-0003-0924-0118'
+    affiliation: 'University of Antwerp, Belgium'
+  - given-names: Hannes
+    family-names: Svardal
+    affiliation: >-
+      University of Antwerp, Belgium and Naturalis
+      Biodiversity Center, Leiden, Netherlands
+    orcid: 'https://orcid.org/0000-0001-7866-7313'
+title:currocam/IdentityByDescentDispersal.jl: v1.0.0
+version: v1.0.0
+date-released: 2026-02-13
+                                
