### 2026-02-13 b24ac93
Fix CITATION.cff
