+++ v1.0.0/inference.jl
@@ -1,0 +1,258 @@
+# # Inference and model evaluation
+# In this section, we go through the process of estimating the parameters of the model
+# and evaluating the fitted model.
+#
+# ## Simulating synthetic datasets
+# The method proposed by [ringbauer_inferring_2017](@cite) makes some assumptions to approximate the evolution
+# of a population in a continuous space. We refer to the [Theory overview](@ref) for an introduction
+# to the underlying theory. In any case, we advise researchers to simulate synthetic data that resembles
+# the empirical data according to their expertise. This is a crucial step to determine the expected
+# performance of the method and to guide different steps of the pre-processing strategy (such as which software
+# to use to detect identity-by-descent blocks or how to bin the observed data).
+#
+# For this purpose, we provide a `SLiM` [haller_tree-sequence_2019](@cite) model to simulate an individual-based population in a continuous
+# space as well as two scripts that fit the model with either error-free identity-by-descent blocks
+# or blocks detected using state-of-the-art software and compare it with the ground-truth parameters.
+# Such models can be found at [simulations directory](https://github.com/currocam/IdentityByDescentDispersal.jl/tree/main/simulations) and might be adapted to your specific needs.
+#
+# ## Processing pipeline
+# Inferring the effective density and dispersal rate from empirical datasets requires running an identity-by-descent detection
+# software across different phased VCF files, post-processing them, and aggregating them appropriately. This
+# procedure is error-prone and time-consuming.
+#
+# To facilitate the analysis, we provide a [Snakemake pipeline](https://github.com/currocam/IdentityByDescentDispersal.jl/blob/main/Snakefile) that can be used to perform a complete analysis,
+# from detecting IBD blocks using [HapIBD](https://github.com/browning-lab/hap-ibd), post-processing them with [Refined IBD](https://faculty.washington.edu/browning/refined-ibd.html), producing a CSV directly compatible
+# with this package, and, optionally, finding a preliminary maximum likelihood estimate of the effective density and effective dispersal rate.
+# [Snakemake](https://snakemake.readthedocs.io/) is a popular tool in bioinformatics
+# and it allows easy parallelization across multiple cores or jobs when submitting to a cluster via SLURM, as well as
+# automatically managing dependencies via Conda or Apptainer containers. Most of the time, running such a pipeline
+# would require a single-command such as:
+# ``` bash
+# snakemake -s Snakefile --configfile config.yaml --cores 8 --sdm conda
+# ```
+# The pipeline is configured via a configuration YAML file. An example of such a configuration file can be found [here](https://github.com/currocam/IdentityByDescentDispersal.jl/blob/main/.test-workflow/config.yaml). You can also find
+# a toy compatible dataset at the [.test-workflow subdirectory](https://github.com/currocam/IdentityByDescentDispersal.jl/tree/main/.test-workflow)
+# We refer to the [extensive documentation of Snakemake](https://snakemake.readthedocs.io/) for more information.
+#
+# ## Model inference
+# Next, we exemplify the common steps of an analysis using `IdentityByDescentDispersal`. First, we load a synthetic
+# dataset for which we know the ground truth. This is the dataset we simulated when showcasing how to [simulate synthetic
+# datasets using SLiM](https://github.com/currocam/IdentityByDescentDispersal.jl/blob/main/simulations/simulate_constant_density_ground_truth.md).
+using JLD2, DataFrames, IdentityByDescentDispersal, Turing, Random
+Random.seed!(1234)
+# Load the synthetic dataset from the package's docs/data directory
+data = load(
[…]
