+++ v1.0.0/tutorial.jl
@@ -1,0 +1,246 @@
+# # Basic usage
+#
+# This section demonstrates the basic usage of the package. We refer to the [Inference and model evaluation](@ref) section
+# of the documentation for an explanation of how to use this package to estimate parameters of demographic models from observed data.
+
+using IdentityByDescentDispersal
+
+# ## Expected density of identity-by-descent blocks
+
+# Genetic distances between populations are often a function of geographic distance. This phenomenon is often referred to as isolation-by-distance.
+# Isolation-by-distance is reflected in the patterns of identity-by-descent blocks (also known as IBD blocks).
+# It is possible to derive the expected patterns of shared identity-by-descent blocks under a diffusion approximation and certain demographic models.
+# We refer to the publication of [ringbauer_inferring_2017](@cite) and the [Theory overview](@ref) section of the documentation for more details.
+#
+# The main feature of this package is to compute the expected density of identity-by-descent blocks under different demographic scenarios.
+# Next, we show how to compute the expected density of identity-by-descent blocks under a constant population density using the `expected_ibd_blocks_constant_density` function.
+
+using Plots
+using QuadGK
+
+L = 0.01         # Block length threshold (in Morgans)
+G = 1.0          # Genome length (in Morgans)
+r_values = range(0.01, 25.0, length = 200);  # Distances
+
+plot(
+    r_values,
+    expected_ibd_blocks_constant_density.(r_values, 1.0, 0.5, L, G),
+    xlabel = "Geographic distance between individuals (r)",
+    ylabel = "E[#IBD blocks per unit of block length and per pair]",
+    title = "Constant effective density scenario",
+    label = "D=1.0, σ=0.5",
+)
+plot!(
+    r_values,
+    expected_ibd_blocks_constant_density.(r_values, 2.0, 0.5, L, G),
+    label = "D=2.0, σ=0.5",
+)
+plot!(
+    r_values,
+    expected_ibd_blocks_constant_density.(r_values, 1.0, 0.8, L, G),
+    label = "D=1.0, σ=0.8",
+)
+
+# The figure above shows how the density of IBD blocks of length 1 centimorgan in a genome of 1 Morgan decays as a function of geographic distance
+# for different demographic scenarios that vary in effective density and dispersal rate.
+# The effective density is in units of number of diploid individuals per area unit (e.g., km²). The dispersal rate is defined as the square root
+# of the average squared axial parent-offspring distance. The rate of the decay of identity-by-descent blocks is directly related
+# to the time since the most recent common ancestor.
+#
+# Recall that the length of IBD blocks is a continuous random variable. Therefore, the expected number of IBD blocks of exactly length `L` is zero.
+# What the `expected_ibd_blocks_constant_density` function returns is the expected number of IBD blocks of length `L` per unit of block
+# length (in Morgans) and per pair. In order to compare such expectations with observed data, we often want to consider instead the
+# expected number of blocks whose length falls in a given interval `[a, b]`. For a small enough interval $[L, L + \Delta L]$ we can approximate
+# the expected number of blocks whose length falls in that interval as:
+#
+# $E[N_{[L, L + \Delta L]}] = \int_{L}^{L + \Delta L} E[N_L] dL \approx E[N_L] \Delta L$
+#
+# For example, the expected number of identity-by-descent blocks whose length falls in the interval `[1cM, 1.1cM]` shared by two diploid
[…]
