+++ versions.js
@@ -1,7 +1,8 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.0",
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "0.1.4";
+var DOCUMENTER_NEWEST = "v1.0.0";
 var DOCUMENTER_STABLE = "stable";

[…]
