### 2026-02-13 #9517 @currocam
@espottesmith I think I've finished.

- [x] Double-check authors and affiliations (including ORCIDs)
- [x] I've released `v1.0.0+docs`
- [x] I've archived the repository in Zenodo, with [DOI: 10.5281/zenodo.18630569](https://doi.org/10.5281/zenodo.18630569)
- [x] I've added a `CITATION.cff` that matches the JOSS paper authorship.
- [x] I've checked the licence (MIT) is properly listed
