### 2026-02-16 #9517 @espottesmith
@currocam some notes:

- The title of the Zenodo archive should exactly match your paper
- Version should exactly match the paper version (I see you have 1.0.0 and 1.0.0+docs; which is the correct version)
