### 2026-02-16 #9517 @espottesmith
Please confirm that the two skipped references do not have DOIs associated with them
