### 2026-02-16 #9517 @currocam
> [@currocam](https://github.com/currocam) some notes:
>
> * The title of the Zenodo archive should exactly match your paper
> * Version should exactly match the paper version (I see you have 1.0.0 and 1.0.0+docs; which is the correct version)

Thanks @espottesmith,
Sorry for the mistake. I have updated the Zenodo archive to match exactly the paper and edited my previous comment to say I released version `v1.0.0+docs`

> Please confirm that the two skipped references do not have DOIs associated with them

For `QuadGK.jl`,  I took the BIB file from https://github.com/JuliaMath/QuadGK.jl/blob/master/CITATION.bib
And I can confirm it has no DOI.

For Turing.jl, I can see that a few months ago they updated the BIB file to include a more recent paper that does have a DOI. Should I update the reference to point to the latest package?
https://github.com/TuringLang/Turing.jl/blob/main/CITATION.bib
