### 2026-02-16 #9517 @espottesmith
If such a citation would be appropriate (i.e., if your package can use the newer version of Turing.jl), then yes, please do update to use the more recent publication.
