### 2026-02-16 2fd8e0a
Update Turing citation
