+++ docs/src/refs.bib
@@ -243,18 +243,15 @@
 	pages = {65--98},
 }
 
-@inproceedings{ge_turing_2018,
-	title = {Turing: {A} {Language} for {Flexible} {Probabilistic} {Inference}},
-	shorttitle = {Turing},
-	url = {https://proceedings.mlr.press/v84/ge18b.html},
-	abstract = {Probabilistic programming promises to simplify and democratize probabilistic machine learning, but successful probabilistic programming systems require flexible, generic and efficient inference engines. In this work, we present a system called Turing for building MCMC algorithms for probabilistic programming inference. Turing has a very simple syntax and makes full use of the numerical capabilities in the Julia programming language, including all implemented probability distributions, and automatic differentiation. Turing supports a wide range of popular Monte Carlo algorithms, including Hamiltonian Monte Carlo (HMC), HMC with No-U-Turns (NUTS), Gibbs sampling, sequential Monte Carlo (SMC), and several particle MCMC (PMCMC) samplers. Most importantly, Turing inference is composable: it combines MCMC operations on subsets of variables, for example using a combination of an HMC engine and a particle Gibbs (PG) engine.  We explore several combinations of inference methods with the aim of finding approaches that are both efficient and universal, i.e. applicable to arbitrary probabilistic models. NUTS—a popular variant of HMC that adapts Hamiltonian simulation path length automatically, although quite powerful for exploring differentiable target distributions, is however not universal. We identify some failure modes for the NUTS engine, and demonstrate that composition of PG (for discrete variables) and NUTS (for continuous variables) can be useful when the NUTS engine is either not applicable, or simply does not work well. Our aim is to present Turing and its composable inference engines to the world and encourage other researchers to build on this system to help advance the field of probabilistic machine learning.},
-	language = {en},
-	urldate = {2025-06-30},
-	booktitle = {Proceedings of the {Twenty}-{First} {International} {Conference} on {Artificial} {Intelligence} and {Statistics}},
-	publisher = {PMLR},
-	author = {Ge, Hong and Xu, Kai and Ghahramani, Zoubin},
-	month = mar,
-	year = {2018},
-	note = {ISSN: 2640-3498},
-	pages = {1682--1690},
+@article{turing,
+author = {Fjelde, Tor Erlend and Xu, Kai and Widmann, David and Tarek, Mohamed and Pfiffer, Cameron and Trapp, Martin and Axen, Seth D. and Sun, Xianda and Hauru, Markus and Yong, Penelope and Tebbutt, Will and Ghahramani, Zoubin and Ge, Hong},
+title = {Turing.jl: a general-purpose probabilistic programming language},
+year = {2025},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3711897},
+doi = {10.1145/3711897},
+note = {Just Accepted},
+journal = {ACM Trans. Probab. Mach. Learn.},
+month = feb,
 }
