### 2026-02-16 #9517 @currocam
> If such a citation would be appropriate (i.e., if your package can use the newer version of Turing.jl), then yes, please do update to use the more recent publication.

I don’t think there’s a “new” Turing.jl, but I’ve updated the citation. They surely prefer citations to go to the new manuscript.
