### 2026-02-16 11c12ec
Revert "Update Turing citation"
This reverts commit 2fd8e0a67e6202cd49dc51e767342f86876e48a6.
