### 2026-02-16 40de7bc
Update Turing citation JOSS draft
