### 2026-02-17 #9517 @espottesmith
Comments on your manuscript:

- "Despite their promising results,22 there is to this date no general-purpose software implementation of their method" <-- what does "general-purpose" mean in this context? This is a pretty specified method, no?
-  "together95 with the popular Turing.jl framework (Fjelde et al., 2025) using a dataset we simulate in96 the documentation" <-- this statement would benefit from a (very brief) mention of what Turing.jl does/is popular for (i.e., Bayesian inference).

Otherwise, the manuscript looks pretty good to me!
