### 2026-02-17 #9517 @currocam
> Comments on your manuscript:
>
> * "Despite their promising results,22 there is to this date no general-purpose software implementation of their method" <-- what does "general-purpose" mean in this context? This is a pretty specified method, no?

Thanks for pointing it out. What we meant by “general-purpose software implementation of their method” is that it’s designed to be used by others in any sort of analysis. Ringbauer et al. released the [code](https://git.ista.ac.at/harald.ringbauer/IBD-Analysis/-/blob/master/POPRES-Analysis/main.py) they used for their specific analysis, but I would say it’s not “general”, as it’s not designed to be used for other purposes (this is not a criticism, I think it’s great work).

In fact, the initial motivation behind writing this library is that we couldn’t figure out how to adapt their scripts. Apparently, this is also a shared experience by another research group who contacted us when we published the preprint to let us know they were using `IdentityByDescentDispersal.jl`.

I think that, in addition to other technical improvements we mention in the manuscript, this adds value to the software and we wanted to reflect that. That said, I agree that “general-purpose software implementation of their method” might not be clear enough and we are happy to rephrase.

> * "together95 with the popular Turing.jl framework (Fjelde et al., 2025) using a dataset we simulate in96 the documentation" <-- this statement would benefit from a (very brief) mention of what Turing.jl does/is popular for (i.e., Bayesian inference).
>
> Otherwise, the manuscript looks pretty good to me!

Great! I will re-rephrase it.
