### 2026-02-18 #9517 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
