### 2026-02-19 9a26734
Fix author's last name formatting
