### 2026-03-06 #9517 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologize for the delay in getting to this step!). Below are some final checks, some of which may require your attention:
### References/Citations
- [x] Please ensure all journal titles are in sentence case
- [x] Geoga et al. arXiv citation can be replaced with peer-reviewed publication: https://link.springer.com/article/10.1007/s11222-022-10127-w
### Copyediting
- [x] line 49: hyphenate "machine-learning" as it is a [compound adjective](https://www.grammarly.com/blog/parts-of-speech/compound-adjectives/) here
- [x] line 53: either consistently `monospace` software titles (e.g. disperseNN2) or don't
- [x] 73: "power density" doesn't need to be hyphenated
- [x] line 87: missing a serial comma after "genetic maps"
- [x] caption of figure 1: missing a noun – we don't expect _what_ to be well-calibrated? (I assume just "the posterior" or "the result" or something would suffice here...)
