### 2026-03-09 7c34a22
journal titles are in sentence case
