+++ manuscript/paper.bib
@@ -26,7 +26,7 @@
 	abstract = {The SLiM software framework for genetically explicit forward simulation has been widely used in population genetics. However, it has been largely restricted to modeling only a single species, which has limited its broader utility in evolutionary biology. Indeed, to our knowledge no general-purpose, flexible modeling framework exists that provides support for simulating multiple species while also providing other key features, such as explicit genetics and continuous space. The lack of such software has limited our ability to model higher biological levels such as communities, ecosystems, coevolutionary and eco-evolutionary processes, and biodiversity, which is crucial for many purposes, from extending our basic understanding of evolutionary ecology to informing conservation and management decisions. We here announce the release of SLiM 4, which fills this important gap by adding support for multiple species, including ecological interactions between species such as predation, parasitism, and mutualism, and illustrate its new features with examples.},
 	number = {5},
 	urldate = {2024-12-05},
-	journal = {The American Naturalist},
+	journal = {The American naturalist},
 	author = {Haller, Benjamin C. and Messer, Philipp W.},
 	month = may,
 	year = {2023},
@@ -99,7 +99,7 @@
 	language = {en},
 	number = {1},
 	urldate = {2025-06-27},
-	journal = {BMC Bioinformatics},
+	journal = {BMC bioinformatics},
 	author = {Smith, Chris C. R. and Kern, Andrew D.},
 	month = oct,
 	year = {2023},
@@ -165,7 +165,7 @@
 	doi = {10.1146/annurev-genet-110711-155534},
 	abstract = {Short segments of identity by descent (IBD) between individuals with no known relationship can be detected using genome-wide single nucleotide polymorphism data and recently developed statistical methodology. Emerging applications for the detected IBD segments include IBD mapping, haplotype phase inference, genotype imputation, and inference of population structure. In this review, we explain the principles behind methods for IBD segment detection, describe recently developed methods, discuss approaches to comparing methods, and give an overview of applications.},
 	language = {eng},
-	journal = {Annual Review of Genetics},
+	journal = {Annual review of genetics},
 	author = {Browning, Sharon R. and Browning, Brian L.},
 	year = {2012},
 	pmid = {22994355},
@@ -197,7 +197,7 @@
 	abstract = {Segments of identity by descent (IBD) are used in many genetic analyses. We present a method for detecting identical-by-descent haplotype segments in phased genotype data. Our method, called hap-IBD, combines a compressed representation of haplotype data, the positional Burrows-Wheeler transform, and multi-threaded execution to produce very fast analysis times. An attractive feature of hap-IBD is its simplicity: the input parameters clearly and precisely define the IBD segments that are reported, so that program correctness can be confirmed by users. We evaluate hap-IBD and four state-of-the-art IBD segment detection methods (GERMLINE, iLASH, RaPID, and TRUFFLE) using UK Biobank chromosome 20 data and simulated sequence data. We show that hap-IBD detects IBD segments faster and more accurately than competing methods, and that hap-IBD is the only method that can rapidly and accurately detect short 2-4 centiMorgan (cM) IBD segments in the full UK Biobank data. Analysis of 485,346 UK Biobank samples through the use of hap-IBD with 12 computational threads detects 231.5 billion autosomal IBD segments with length ≥2 cM in 24.4 h.},
 	language = {eng},
 	number = {4},
-	journal = {American Journal of Human Genetics},
+	journal = {American journal of human genetics},
 	author = {Zhou, Ying and Browning, Sharon R. and Browning, Brian L.},
 	month = apr,
 	year = {2020},
@@ -235,7 +235,7 @@
[…]
