+++ manuscript/paper.md
@@ -44,7 +44,7 @@ The distribution of recent coalescent events between individuals in space can be
 
 In order to make the inference approach available to the broader audience of evolutionary biologists and conservation scientists, we present `IdentityByDescentDispersal.jl`, a Julia [@bezanson_julia_2017] package with an efficient and easy-to-use implementation of the method. The package implements the core equations proposed by @ringbauer_inferring_2017 and can be used to perform composite likelihood-based inference using either maximum-likelihood estimation (MLE) or Bayesian inference.
 
-The method of @ringbauer_inferring_2017 was limited to a family of functions for the change in effective population density over time of the form $D_e(t) = Dt^{-\beta}$, for which the theory was analytically tractable. In addition, in the paper describing the original approach, the authors used gradient-free optimization to calculate maximum likelihood estimates (MLEs). Our implementation makes two major software contributions. First, we admit composite likelihood calculations for arbitrary functions $D_e(t)$ by evaluating the relevant integrals numerically through Gaussian quadrature rules [@quadgk]. This significantly enlarges the space of biologically relevant models that can be fitted. Second, our implementation takes advantage of the powerful Julia ecosystem and the work of @geoga_fitting_2022 to provide a version of the composite likelihood that is fully compatible with automatic differentiation (AD), including AD with respect to $\beta$. By having a fully AD-compatible composite likelihood, `IdentityByDescentDispersal.jl` can be used together with standard gradient-based optimization and sampling methods available in the Julia ecosystem, which are typically more efficient than gradient-free methods.
+The method of @ringbauer_inferring_2017 was limited to a family of functions for the change in effective population density over time of the form $D_e(t) = Dt^{-\beta}$, for which the theory was analytically tractable. In addition, in the paper describing the original approach, the authors used gradient-free optimization to calculate maximum likelihood estimates (MLEs). Our implementation makes two major software contributions. First, we admit composite likelihood calculations for arbitrary functions $D_e(t)$ by evaluating the relevant integrals numerically through Gaussian quadrature rules [@quadgk]. This significantly enlarges the space of biologically relevant models that can be fitted. Second, our implementation takes advantage of the powerful Julia ecosystem and the work of @Geoga_2023 to provide a version of the composite likelihood that is fully compatible with automatic differentiation (AD), including AD with respect to $\beta$. By having a fully AD-compatible composite likelihood, `IdentityByDescentDispersal.jl` can be used together with standard gradient-based optimization and sampling methods available in the Julia ecosystem, which are typically more efficient than gradient-free methods.
 
 Lastly, our package comes with a template to simulate synthetic datasets and a pipeline for end-to-end analysis from VCF files to final estimates. We believe it will encourage a broader audience to adopt the inference scheme proposed by @ringbauer_inferring_2017, motivate further developments and expand its applications.
 
