+++ manuscript/paper.md
@@ -36,34 +36,34 @@ bibliography: paper.bib
 
 The population density and per-generation dispersal rate of a population are central parameters in the study of evolution and ecology. The dispersal rate is particularly relevant for conservation management of fragmented or invasive species [@driscoll_trajectory_2014]. There is a growing interest in developing statistical methods that exploit the increasingly available genetic data to estimate the effective population density and effective dispersal rate [@rousset_genetic_1997;@ringbauer_inferring_2017;@smith_dispersal_2023;@smith_dispersenn2_2023].
 
-The distribution of recent coalescent events between individuals in space can be used to estimate such quantities through the distribution of identity-by-descent (IBD) blocks [@ringbauer_inferring_2017]. An IBD block is defined as a segment of DNA that has been inherited by a pair of individuals from a common ancestor without being broken by recombination. Here we present `IdentityByDescentDispersal.jl`, a Julia package for estimating effective population densities and dispersal rates from observed spatial patterns of IBD shared blocks.
+The distribution of recent coalescent events between individuals in space can be used to estimate such quantities through the distribution of identity-by-descent (IBD) blocks [@ringbauer_inferring_2017]. An IBD block is defined as a segment of DNA that has been inherited by a pair of individuals from a common ancestor without being broken by recombination. Here we present IdentityByDescentDispersal.jl, a Julia package for estimating effective population densities and dispersal rates from observed spatial patterns of IBD shared blocks.
 
 # Statement of need
 
 @ringbauer_inferring_2017 proposed an inference scheme for the estimation of effective population density and effective dispersal rate from shared IBD blocks. Despite their promising results, there is to this date no general-purpose software implementation of their method.
 
-In order to make the inference approach available to the broader audience of evolutionary biologists and conservation scientists, we present `IdentityByDescentDispersal.jl`, a Julia [@bezanson_julia_2017] package with an efficient and easy-to-use implementation of the method. The package implements the core equations proposed by @ringbauer_inferring_2017 and can be used to perform composite likelihood-based inference using either maximum-likelihood estimation (MLE) or Bayesian inference.
+In order to make the inference approach available to the broader audience of evolutionary biologists and conservation scientists, we present IdentityByDescentDispersal.jl, a Julia [@bezanson_julia_2017] package with an efficient and easy-to-use implementation of the method. The package implements the core equations proposed by @ringbauer_inferring_2017 and can be used to perform composite likelihood-based inference using either maximum-likelihood estimation (MLE) or Bayesian inference.
 
[…]
