+++ manuscript/paper.bib
@@ -252,7 +252,7 @@ address = {New York, NY, USA},
 url = {https://doi.org/10.1145/3711897},
 doi = {10.1145/3711897},
 note = {Just Accepted},
-journal = {ACM trans. probab. mach. learn.},
+journal = {ACM transactions on probabilistic machine learning},
 month = feb,
 }
 @misc{quadgk,
