### 2026-03-09 #9517 @currocam
Thanks a lot @rkurchin! No need to apologize for the delay. I really appreciate the time you put into this. I misunderstood the stage of the manuscript. Sorry if I rushed you into looking at this.

About the references, I'm a bit confused with the journal titles. I think I updated the [.bib file](https://github.com/currocam/IdentityByDescentDispersal.jl/blob/76cbe11b6f122864bf82e45527b73a316dc8ab3e/manuscript/paper.bib) correctly, but the PDF doesn't show journal titles in sentence case. Could you guide me on this?
