### 2026-03-15 #9517 @Kevin-Mattheus-Moerman
@currocam congratulations on this JOSS publication! :tada:

At this point I would like to thank @espottesmith for editing this submission.

A special thank you also to the reviewers:   @slwu89, @g-rppl !!!
